+++ territory/src/main/java/pt/estga/territory/services/AdministrativeDivisionService.java
@@ -1,15 +1,10 @@
 package pt.estga.territory.services;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import pt.estga.territory.entities.AdministrativeDivision;
 import pt.estga.territory.repositories.AdministrativeDivisionRepository;
 
-import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -17,14 +12,6 @@ import java.util.Optional;
 public class AdministrativeDivisionService {
 
     private final AdministrativeDivisionRepository repository;
-
-    public Page<AdministrativeDivision> search(Specification<AdministrativeDivision> specification, Pageable pageable) {
-        return repository.findAll(specification, pageable);
-    }
-
-    public Optional<AdministrativeDivision> findById(Long id) {
-        return repository.findById(id);
-    }
 
     public AdministrativeDivision create(AdministrativeDivision division) {
         return repository.save(division);
